Add a first sketch of an alert database server
Add a FastAPI-based server which provides alerts and schemas according
to the routes laid out in DMTN-183.

This is done in a package structure which is separate from the LSST
stack because I do not expect it to depend upon the stack, and the
stack's complexity is not warranted here.